requirements.txt
#--------Flask---------------------
click==6.7
Flask==0.12
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe
Werkzeug
WTForms
wsgiref
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Login
psycopg2
pyjwt
#----------------------------------------
#-----------Production Server------------
gunicorn==19.6.0
#-----------------------------